rename: Alert 관련 테스트 위치 변경 (#196)
